Rename appconfiguration directory => appconfig
