+++ sdk/appconfig/azappconfig/ci.yml
@@ -8,7 +8,7 @@ trigger:
       - release/*
   paths:
     include:
-      - sdk/appconfiguration/azappconfig
+      - sdk/appconfig/azappconfig
 
 pr:
   branches:
@@ -19,10 +19,10 @@ pr:
       - release/*
   paths:
     include:
-      - sdk/appconfiguration/azappconfig
+      - sdk/appconfig/azappconfig
 
 stages:
   - template: /eng/pipelines/templates/jobs/archetype-sdk-client.yml
     parameters:
-      ServiceDirectory: 'appconfiguration/azappconfig'
+      ServiceDirectory: 'appconfig/azappconfig'
       RunLiveTests: true
